Included in this commit:
2 new orbital defence pod models; these are for the Sith Empire. The first is the first stage available to the Sith of this satellite; mounted with two heavy lasers for anti-starfighter duty, the latter is mounted with two light turbolasers; a second stage which increases its potency slightly. The Sith have somewhat more limited orbital defences compared to the Republic in this era, as they are essentially more aggressive.

Both models use the same two textures (with their own normal maps); a gun texture (it is a separate animated mesh, will be coded in to function since they are sole weapons) and the body texture.

New particles: new light blinks for stations, ships or whatever anyone so pleases: red and green blinks are available in different sizes; red and green standard, red 120% size, red and green 300% size. These are used for the starboard and port lamps respectively on the satellites in this commit.

LTCC